Mounter interface should not re-instantiate SafeFormatAndMount
Allowing SafeFormatAndMount to be instantiated multiple times prevents
faking the commands efficiently in unit tests because it wipes out any
exec.Interface fakes that unit tests might have configured.

There is no point in allowing SafeFormatAndMount to be re-instantiated
again by NewSafeFormatAndMount() of the current Mounter interface. It
should be instantiated only once in mountmanager.Mounter constructors.

We can simply replace it with a getter to prevent wiping fakes.

Signed-off-by: Roman Bednar <rbednar@redhat.com>