add new node mounter that allows faking commands
If we need to test any code that executes commands where we need to
parse command output successfully, the FakeExec with
`DisableScripts: true` is not sufficient as it always returns empty
byteslice.

To add more control over command fakes the Mounter should support
injecting CommandScript with a sequence of fake commands/actions.

Signed-off-by: Roman Bednar <rbednar@redhat.com>