allow command faking in mounter (#69)
* Mounter interface should not re-instantiate SafeFormatAndMount

Allowing SafeFormatAndMount to be instantiated multiple times prevents
faking the commands efficiently in unit tests because it wipes out any
exec.Interface fakes that unit tests might have configured.

There is no point in allowing SafeFormatAndMount to be re-instantiated
again by NewSafeFormatAndMount() of the current Mounter interface. It
should be instantiated only once in mountmanager.Mounter constructors.

We can simply replace it with a getter to prevent wiping fakes.

Signed-off-by: Roman Bednar <rbednar@redhat.com>

* add new node mounter that allows faking commands

If we need to test any code that executes commands where we need to
parse command output successfully, the FakeExec with
`DisableScripts: true` is not sufficient as it always returns empty
byteslice.

To add more control over command fakes the Mounter should support
injecting CommandScript with a sequence of fake commands/actions.

Signed-off-by: Roman Bednar <rbednar@redhat.com>

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Signed-off-by: Sameer Shaikh <sameer.shaikh@ibm.com>